server/rpc/requirements.txt
grpcio==1.30.0
grpcio-tools==1.30.0
SQLAlchemy==1.3.24
mysqlclient==1.4.6
pycrypto==2.6.1
python-dateutil==2.8.1